Improve test using `not --crash`.
